+++ packages/studio-base/src/index.ts
@@ -30,11 +30,8 @@ export { default as installDevtoolsFormatters } from "./util/installDevtoolsForm
 export { default as overwriteFetch } from "./util/overwriteFetch";
 export { default as waitForFonts } from "./util/waitForFonts";
 export { IdbExtensionLoader } from "./services/IdbExtensionLoader";
-export type {
-  ExtensionInfo,
-  ExtensionLoader,
-  ExtensionNamespace,
-} from "./services/ExtensionLoader";
+export type { ExtensionLoader } from "./services/ExtensionLoader";
+export type { ExtensionInfo, ExtensionNamespace } from "./types/Extensions";
 export { AppSetting } from "./AppSetting";
 export { default as FoxgloveDataPlatformDataSourceFactory } from "./dataSources/FoxgloveDataPlatformDataSourceFactory";
 export { default as FoxgloveWebSocketDataSourceFactory } from "./dataSources/FoxgloveWebSocketDataSourceFactory";
